kimi_k3: the pipeline integration cells run in the B200 suite
The features suite runs its real-PG cells on every pull request on 8 x A10G,
where Attention Gym's KDA kernel refuses to run. Kimi K3's other integration
cell already lives in the B200 suite for that reason; the two pipeline cells
and their recipes move there too.
