+++ tests/integration_tests/features.py
@@ -320,20 +320,4 @@ def build_features_test_list() -> list[OverrideDefinitions]:
             timeout=30,
             use_real_pg=True,
         ),
-        OverrideDefinitions(
-            configs=[recipes.kimi_k3_debugmodel_pp2_vp2],
-            test_descr="Kimi K3, pipeline parallel pp2 x vp2",
-            test_name="kimi_k3_pp2_vp2",
-            ngpu=2,
-            use_real_pg=True,
-            skip_rocm_test=True,
-        ),
-        OverrideDefinitions(
-            configs=[recipes.kimi_k3_debugmodel_pp8_vp4],
-            test_descr="Kimi K3, pipeline parallel pp8 x vp4",
-            test_name="kimi_k3_pp8_vp4",
-            ngpu=8,
-            use_real_pg=True,
-            skip_rocm_test=True,
-        ),
     ]
